(feat) handle helm delete hook
When an helm chart is removed in pull mode:
1. if delete hooks are present, pass those resources to the applier
2. wait for the applier to provision those
3. proceed removing all resources

When helm resources are collected and made available for sveltos-applier
the order is following:

1. pre-install/pre-upgrade hook resources
2. CRDs
3. other resources
4. post-install/post-upgrade hook resources

When helm is uninstalled only delete hook resources are passed.
First pre-delete then the post-delete.

In order to detect if an operation is install/upgrade/rollback when
cluster is in pull mode, Sveltos keeps tracks in memory (chart-manager)
the versions current deployed in each cluster.
This in memory map is automatically rebuilt on restart by looking at
the ClusterSummary Spec (to get version) and Status (to detect which ClusterSummary
is managing an helm chart)

+++ cmd/main.go
@@ -329,7 +329,10 @@ func setupChecks(mgr ctrl.Manager) {
 }
 
 // capiCRDHandler restarts process if a CAPI CRD is updated
-func capiCRDHandler(gvk *schema.GroupVersionKind) {
+func capiCRDHandler(gvk *schema.GroupVersionKind, action crd.ChangeType) {
+	if action == crd.Modify {
+		return
+	}
 	if gvk.Group == clusterv1.GroupVersion.Group {
 		if killErr := syscall.Kill(syscall.Getpid(), syscall.SIGTERM); killErr != nil {
 			panic("kill -TERM failed")
@@ -353,7 +356,10 @@ func isCAPIInstalled(ctx context.Context, c client.Client) (bool, error) {
 }
 
 // fluxCRDHandler restarts process if a Flux CRD is updated
-func fluxCRDHandler(gvk *schema.GroupVersionKind) {
+func fluxCRDHandler(gvk *schema.GroupVersionKind, action crd.ChangeType) {
+	if action == crd.Modify {
+		return
+	}
 	if gvk.Group == sourcev1.GroupVersion.Group {
 		if killErr := syscall.Kill(syscall.Getpid(), syscall.SIGTERM); killErr != nil {
 			panic("kill -TERM failed")
